project-1/requirements.txt
torch
torchvision
torchaudio
torchtext
streamlit
ollama
opencv-python
transformers
dorothy-cci
datasets 
tiktoken 
onnxruntime
Pillow
imageio
sounddevice
librosa
ultralytics
scikit-learn
scikit-image
acids-rave
open_clip_torch
streamlit-drawable-canvas
Levenshtein